Moved query related types and functions to the dedicated folder
