+++ src/packages/dumbo/src/core/execute/execute.ts
@@ -1,5 +1,4 @@
 import type { Connection } from '../connections';
-import type { QueryResult, QueryResultRow } from '../connections/execute';
 
 export const execute = async <
   Result = void,
@@ -88,86 +87,3 @@ export const execute = async <
 
 //     return { success: true, result: undefined };
 //   });
-
-export const firstOrNull = async <
-  Result extends QueryResultRow = QueryResultRow,
->(
-  getResult: Promise<QueryResult<Result>>,
-): Promise<Result | null> => {
-  const result = await getResult;
-
-  return result.rows.length > 0 ? result.rows[0] ?? null : null;
-};
-
-export const first = async <Result extends QueryResultRow = QueryResultRow>(
-  getResult: Promise<QueryResult<Result>>,
-): Promise<Result> => {
-  const result = await getResult;
-
-  if (result.rows.length === 0)
-    throw new Error("Query didn't return any result");
-
-  return result.rows[0]!;
-};
-
-export const singleOrNull = async <
-  Result extends QueryResultRow = QueryResultRow,
->(
-  getResult: Promise<QueryResult<Result>>,
-): Promise<Result | null> => {
-  const result = await getResult;
-
-  if (result.rows.length > 1) throw new Error('Query had more than one result');
-
-  return result.rows.length > 0 ? result.rows[0] ?? null : null;
-};
-
-export const single = async <Result extends QueryResultRow = QueryResultRow>(
-  getResult: Promise<QueryResult<Result>>,
-): Promise<Result> => {
-  const result = await getResult;
-
-  if (result.rows.length === 0)
-    throw new Error("Query didn't return any result");
-
-  if (result.rows.length > 1) throw new Error('Query had more than one result');
-
-  return result.rows[0]!;
-};
-
-export type ExistsSQLQueryResult = { exists: boolean };
-
-export const exists = async (
-  getResult: Promise<QueryResult<ExistsSQLQueryResult>>,
-): Promise<boolean> => {
-  const result = await single(getResult);
-
-  return result.exists === true;
-};
-
-export const mapRows = async <
-  Result extends QueryResultRow = QueryResultRow,
-  Mapped = unknown,
->(
-  getResult: Promise<QueryResult<Result>>,
-  map: (row: Result) => Mapped,
-): Promise<Mapped[]> => {
-  const result = await getResult;
-
-  return result.rows.map(map);
-};
-
-export const toCamelCase = (snakeStr: string): string =>
-  snakeStr.replace(/_([a-z])/g, (g) => g[1]?.toUpperCase() ?? '');
-
-export const mapToCamelCase = <T extends Record<string, unknown>>(
-  obj: T,
-): T => {
-  const newObj: Record<string, unknown> = {};
-  for (const key in obj) {
-    if (Object.prototype.hasOwnProperty.call(obj, key)) {
-      newObj[toCamelCase(key)] = obj[key];
-    }
-  }
-  return newObj as T;
-};
