+++ src/packages/dumbo/src/core/index.ts
@@ -1,3 +1,4 @@
 export * from './connections';
 export * from './execute';
+export * from './query';
 export * from './sql';
